add SongPlayer
The gameplay still needs to be implimented but there is now a songPlayer file that can be extended let the user type in characters and have different songs loaded into the text field.
Also some menu functionality was added to let the player navigate from the main menu to the songplayer and navigate out of the game.
navigation from the song player back to the menu or some other way to quit still needs to be implimented.

Closes #1, Closes #2