add wpm mode for cpu cursor
wpm mode is words per minute mode. It's the easiest one to add right now and will be a good default in the short term. This is also required for one of the overall goals of the repo: "to be be a useful tool if people just want to listen to music while reaching a typing speed benchmark"