add another japanese anthem
this one probably is protected by copyright but for adequite testing i thought we would need a second japanese song. If anyone has a better alternative I'd be fine with replacing this. At the very least this seems niche enough that we aren't likely to be sued, especially given our use case.