Create legal notices screen
When ever you distribute a copy of an open-source piece of software, you
have to make sure that you abide by the terms of any licenses that apply
to the software.

Before this change, it was easy to distribute copies of Song Typing’s
source code, but it was hard to distribute binary copies of Song Typing.
As far as I know, Song Typing’s source code already contained all
required legal notices. Binaries of Song Typing on the other had did not
contain all of the required legal notices. If someone wanted to
distribute binaries of Song Typing, then they would have to track down
every required legal notice and add each one to the distribution.

This commit makes it easier to distribute Song Typing binaries by making
sure that binaries contain all legal notices automatically.

---

One of the legal notices that this commit adds is a bit unusual. The
unusual legal notice says:

> {game_name} uses Godot Engine version {godot_version}. If you want the
> source code for that version of the Godot Engine, then please visit
> this page: <[url]https://docs.godotengine.org/en/stable/contributing/development/compiling/getting_source.html[/url]>.
>
> The Godot Engine source code contains a file named
> [code]ca-certificates.crt[/code]. If you want the source code for
> [code]ca-certificates.crt[/code], then please take a look at the
> [code]thirdparty/README.md[/code] file that’s inside Godot’s source
> code, and then take a look at the table that’s on this page:
> <[url]https://github.com/Jayman2000/ca-bundle-pr/tree/corresponding-source-for-old-commits#certdatatxt[/url]>.

That unusual legal notice exists to help work around the fact that Godot
uses an old version of ca-bundle at the moment [1]. Unfortunately, old
versions of ca-bundle have a but that’s related to MPL 2.0 compliance
[2]. That unusual legal notice helps compensate for the bug in old
versions of ca-bundle.

[1]: <https://github.com/godotengine/godot/tree/4.3-stable/thirdparty#certs>
[2]: <https://github.com/bagder/ca-bundle/issues/11>