Prevent game from crashing when a level ends
Before this change, the game would crash when you got to the end of a
level. This change prevents that crash from happening. Instead of
crashing, the game will now leave the last line of lyrics highlighted.
The final line will stay highlighted, but nothing will happen
afterwards. That’s probably not ideal, but it’s better than crashing.

Fixes #50.